+++ src/Plugins/Collidable/index.js
@@ -1,0 +1,3 @@
+import Collidable from './Collidable';
+
+export default Collidable;
