+++ src/Plugins/Snappable/README.md
@@ -1,0 +1,1 @@
+## Snappable
