+++ src/Plugins/Snappable/index.js
@@ -1,0 +1,3 @@
+import Snappable from './Snappable';
+
+export default Snappable;
