+++ src/Plugins/index.js
@@ -1,0 +1,7 @@
+import Collidable from './Collidable';
+import Snappable from './Snappable';
+
+export {
+  Collidable,
+  Snappable,
+};
