+++ src/index.js
@@ -3,9 +3,9 @@ import Droppable from './Droppable';
 import Swappable from './Swappable';
 import Sortable from './Sortable';
 
+import {Snappable, Collidable} from './Plugins';
+
 import AbstractEvent from './events/abstract-event';
-import Snappable from './behaviour/snappable';
-import Collidable from './behaviour/collidable';
 
 export {
   Draggable,
